Only install if not present

+++ work-setup.sh
@@ -2,14 +2,26 @@
 set -euo pipefail
 
 # oh-my-zsh
-rm -rf ~/.oh-my-zsh
-ZSH= sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+if [[ -z ${ZSH+x} ]]; then
+    echo "Installing Oh My Zsh"
+    # TODO: figure out how to prevent this command from exiting the script
+    sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+else
+    echo "Oh My Zsh already present, skipping..."
+fi
 
 # homebrew
 if ! xcode-select -p > /dev/null; then
     xcode-select --install
 fi
-/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+
+if [[ $(command -v brew) == "" ]]; then
+    echo "Installing brew"
+    /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+else
+    echo "brew already present, updating..."
+    brew update && brew upgrade
+fi
 
 # starship
 brew install starship
@@ -22,10 +34,15 @@ git config --global pull.rebase true
 
 # fzf
 brew install fzf
-$(brew --prefix)/opt/fzf/install --completion --no-update-rc
+$(brew --prefix)/opt/fzf/install --key-bindings --completion --no-update-rc
 
 # z
-curl -o ~/z.sh https://raw.githubusercontent.com/rupa/z/master/z.sh
+if [[ ! -f ~/z.sh ]]; then
+    echo "Installing z"
+    curl -o ~/z.sh https://raw.githubusercontent.com/rupa/z/master/z.sh
+else
+    echo "z already present, skipping..."
+fi
 
 # bat
 brew install bat
@@ -34,9 +51,11 @@ brew install bat
 brew install neovim
 
 # phpbrew
+brew install php
+brew install autoconf pkg-config bzip2 zlib openssl@1.1 oniguruma readline libzip libxml2
 curl -L -O https://github.com/phpbrew/phpbrew/releases/latest/download/phpbrew.phar
 chmod +x phpbrew.phar
-sudo mv phpbrew.phar /usr/local/bin/phpbrew
+mv phpbrew.phar /usr/local/bin/phpbrew
 phpbrew init
 
 # pyenv
@@ -48,7 +67,7 @@ brew install pyenv-virtualenvwrapper
 
 # python
 brew install xz
-pyenv install 3.12
+pyenv install --skip-existing 3.12
 pyenv global 3.12
 python -m pip install --upgrade pip setuptools wheel
 
@@ -57,10 +76,10 @@ curl -sSL https://install.python-poetry.org | python3 -
 
 # ruby
 brew install chruby ruby-install
-ruby-install ruby 3.3.0
+ruby-install --no-reinstall ruby 3.3.0
 
 # casks
-brew install --cask iterm2 visual-studio-code pycharm-ce docker rectangle opera
+#brew install --adopt --casks iterm2 visual-studio-code pycharm-ce docker rectangle opera
 
 cp .zshrc ~/.zshrc
 source ~/.zshrc
